Improve installation instructions
`sobelow` is often running in continuous integration pipelines and when
it does, it will more often than not be in the test environment. On top
of this, `sobelow` doesn't have to be started as a [runtime
application](https://hexdocs.pm/mix/Mix.Tasks.Deps.html#module-dependency-definition-options),
so `runtime: false` can be added as well.

Closes #120